requirements.txt
pip==23.3.1
setuptools==69.0.2
wheel==0.42.0
python-dotenv==1.0.0
python-telegram-bot==20.7
